🚨 wire — clear all ruff + mypy debt; re-gate CI (behavior-preserving)
Fix the wire server's pre-existing 53 ruff + 30 mypy issues with no runtime-behavior change; 202 tests unchanged. Re-enable CI gating by dropping the two continue-on-error flags on the wire Lint/Typecheck steps. Notable: F821 was a real latent bug (Presence used in app.py's Message|Presence annotations but never imported); E501 wraps verified byte-identical across all schema/Query/CLI description strings; Config.wait_default/wait_max widened int->float (honest type — only ever asyncio timeouts; zero runtime effect). Verified with the exact CI commands: ruff 0, ruff format clean, mypy 0, pytest 202, tsc 0.

+++ ripperdoc/wetware/wire/server/src/wire/config.py
@@ -57,8 +57,10 @@ class Config:
     secret: str
     topic: str
     public_url: str | None = None
-    wait_default: int = 30
-    wait_max: int = 60
+    # Second-durations used as asyncio.wait_for timeouts, so float-typed; the CLI
+    # still feeds ints (valid floats) and the tests drive sub-second values.
+    wait_default: float = 30
+    wait_max: float = 60
     idle_timeout: int = 300
     sweep_interval: int = 15
     empty_grace: int = 900
